20171130.18:22

+++ build/dev-server.js
@@ -6,7 +6,9 @@ let path = require('path');
 let configBuildPath = path.resolve(__dirname, '../config/index.js');
 let outputPath = path.resolve(configBuildPath, configBuild.dev);
 
-shell.rm('-rf', outputPath);
+if (shell.test('-d', outputPath)) {
+  shell.rm('-rf', outputPath);
+}
 shell.exec('fis3 server stop -p ' + configBuild.port);
 shell.exec('fis3 server start --no-browse --root ' + outputPath + ' -p ' + configBuild.port);
 shell.exec('fis3 release -wLcd ' + outputPath);
